Add disconnect
Allows a user to disconnect the current SSH connection and
run another command. Jsch doesn't allow a disconnect while
it's attempting to connect to a server, so all this really
does is ignore future responses from the plugin. The resources
used by the connection are still allocated until the server
responds with either connection error or success.